+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(rtransparent)
+
+test_check("rtransparent")